:truck: move package to outlier
