+++ src/outlier/__init__.py
@@ -1,0 +1,6 @@
+from outlier import manager
+from outlier import func
+from outlier import manager
+from outlier import protocol
+from outlier import regdecorator
+from outlier import utils
