+++ src/outlier/regdecorator.py
@@ -5,10 +5,10 @@ LastEditTime: 2023-01-10 18:08:13
 FilePath: /outlier/src/regdecorator.py
 '''
 import logging
-from func import FuncBase, State
+from .func import FuncBase, State
 
 import inspect
-# import logging
+
 
 # Server Decorators
