+++ src/outlier/utils.py
@@ -5,6 +5,7 @@ LastEditTime: 2023-01-10 19:58:51
 FilePath: /outlier/src/utils.py
 '''
 import logging
+import socket
 from logging import handlers
 
 FORMAT = '%(asctime)s - %(message)s'
@@ -27,3 +28,17 @@ def retry_process(tryrecall, tryargs, failrecall, failargs, uppertimes, *args):
         except Exception as e:
             failrecall(*failargs)
             errortime += 1
+            
+            
+
+def getConnectAddr() -> str:
+    """This function is to check the ipaddress that a host should be bind
+
+    Returns:
+        str: ip address
+    """
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    s.connect(('8.8.8.8', 80))
+    ip = s.getsockname()
+    s.close()
+    return ip[0]
